Revert "add appbar bottom"
