+++ lib/Library/main_library.dart
@@ -16,7 +16,7 @@ export 'dart:io';
 export 'package:app_wallet/models/expense.dart';
 export 'package:app_wallet/models/category.dart';
 export 'package:app_wallet/models/expense_bucket.dart';
-export 'package:app_wallet/screens/wallet_home_page.dart';
+export 'package:app_wallet/screens/expenses.dart';
 export 'package:app_wallet/screens/new_expense_screen.dart';
 export 'package:app_wallet/screens/log_in.dart';
 export 'package:app_wallet/screens/inicio.dart';
@@ -52,13 +52,3 @@ export 'package:app_wallet/utils/extensions.dart';
 export 'package:app_wallet/components/loaders/wallet_loader.dart';
 export 'package:google_sign_in/google_sign_in.dart';
 export 'package:app_wallet/components/appbar/appbar.dart';
-export 'package:app_wallet/components/appbar/appbar_bottom.dart';
-export 'package:app_wallet/models/category_helper.dart';
-export 'package:app_wallet/components/widgets/calendar_selector.dart';
-export 'package:app_wallet/components/widgets/category_list.dart';
-export 'package:app_wallet/components/widgets/pie_chart.dart';
-export 'package:app_wallet/screens/profile_page.dart';
-export 'package:app_wallet/wallet_controllers/wallet_expenses_controller.dart';
-export 'package:app_wallet/navigation_service/wallet_navigation_service.dart';
-export 'package:app_wallet/components/appbar/wallet_expenses_appbar.dart';
-export 'package:app_wallet/components/filters/wallet_filters_button.dart';
