--- lib/components/appbar/appbar_bottom.dart
@@ -1,83 +1,0 @@
-import 'package:app_wallet/library/main_library.dart';
-
-class WalletBottomAppBar extends StatelessWidget {
-  final int currentIndex;
-  final Function(int) onTap;
-
-  const WalletBottomAppBar({
-    Key? key,
-    required this.currentIndex,
-    required this.onTap,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return BottomAppBar(
-      elevation: 8.0,
-      shape: const CircularNotchedRectangle(),
-      child:  SizedBox(
-        height: 70.0,
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceAround,
-          children: [
-            _buildBottomAppBarItem(
-              icon: Icons.home,
-              label: 'Home',
-              index: 0,
-            ),
-            _buildBottomAppBarItem(
-              icon: Icons.bar_chart,
-              label: 'Estadísticas',
-              index: 1,
-            ),
-            const SizedBox(width: 48), 
-            _buildBottomAppBarItem(
-              icon: Icons.assessment,
-              label: 'Informes',
-              index: 2,
-            ),
-            _buildBottomAppBarItem(
-              icon: Icons.wallet,
-              label: 'MiWallet',
-              index: 3,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildBottomAppBarItem({
-    required IconData icon,
-    required String label,
-    required int index,
-  }) {
-    final bool isSelected = currentIndex == index;
-    
-    return InkWell(
-      onTap: () => onTap(index),
-      child: Container(
-        padding: const EdgeInsets.symmetric(vertical: 6.0), 
-        child: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            Icon(
-              icon,
-              color: isSelected ? AwColors.appBarColor : AwColors.modalGrey,
-              size: AwSize.s22, 
-            ),
-            const SizedBox(height: 2.0), 
-            Text(
-              label,
-              style: TextStyle(
-                fontSize: AwSize.s10, 
-                color: isSelected ? AwColors.appBarColor : AwColors.modalGrey,
-                fontWeight: isSelected ? FontWeight.bold : FontWeight.normal,
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
